@@ -1,5 +1,5 @@
 import React from 'react'
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
+import { BrowserRouter as Router, Routes, Route, useMatch } from 'react-router-dom'
 import { useApi } from './useApi'
 import LoadingSpinner from './LoadingSpinner'
 import ErrorMessage from './ErrorMessage'
@@ -13,7 +13,9 @@ const mapResults = (({ results }) => results.map(({ url, name }) => ({
 })))
 
 const App = () => {
+  const match = useMatch('/pokemon/:name')
   const { data: pokemonList, error, isLoading } = useApi('https://pokeapi.co/api/v2/pokemon/?limit=50', mapResults)
+
   if (isLoading) {
     return <LoadingSpinner />
   }
@@ -21,20 +23,22 @@ const App = () => {
     return <ErrorMessage error={error} />
   }
 
+  let next = null
+  let previous = null
+
+  if (match && match.params) {
+    const pokemonId = pokemonList.find(({ name }) => name === match.params.name).id
+    previous = pokemonList.find(({ id }) => id === pokemonId - 1)
+    next = pokemonList.find(({ id }) => id === pokemonId + 1)
+  }
+
   return (
-    <Router>
-      <Switch>
-        <Route exact path="/">
-          <PokemonList pokemonList={pokemonList} />
-        </Route>
-        <Route path="/pokemon/:name" render={(routeParams) => {
-          const pokemonId = pokemonList.find(({ name }) => name === routeParams.match.params.name).id
-          const previous = pokemonList.find(({ id }) => id === pokemonId - 1)
-          const next = pokemonList.find(({ id }) => id === pokemonId + 1)
-          return <PokemonPage pokemonList={pokemonList} previous={previous} next={next} />
-        }} />
-      </Switch>
-    </Router>
+    <Routes>
+      <Route exact path="/" element={<PokemonList pokemonList={pokemonList} />} />
+      <Route exact path="/pokemon/:name" element={
+        <PokemonPage pokemonList={pokemonList} previous={previous} next={next} />
+      } />
+    </Routes>
   )
 }
 
